Moved GameManagerService to games.guessIt package
